+++ test/server/steps/common.js
@@ -1,0 +1,1 @@
+module.exports = function(){};
