+++ test/server/support/afterEach.js
@@ -1,0 +1,5 @@
+module.exports = function(){
+  this.After(function(){
+    this.server.close();
+  });
+};
